+++ pkgs/checks/test/extensions/function_test.dart
@@ -43,6 +43,20 @@ void main() {
           (Subject<void Function()> it) => it.throws<StateError>(),
         ).description.deepEquals(['  throws an error of type StateError']);
       });
+      test('evaluates condition', () {
+        check(() => throw StateError('oops!')).isRejectedBy(
+          (it) => it.throws<StateError>(
+            (it) => it.has((e) => e.message, 'message').equals('wrong'),
+          ),
+          actual: ["'oops!'"],
+          which: ['differs at offset 0:', '  wrong', '  oops!', '  ^'],
+        );
+      });
+      test('returns valid subject', () {
+        check(
+          () => throw StateError('oops!'),
+        ).throws<StateError>().has((e) => e.message, 'message').equals('oops!');
+      });
     });
 
     group('returnsNormally', () {
@@ -60,6 +74,16 @@ void main() {
           actual: ['a function that throws'],
           which: ['threw <Bad state: oops!> at:', '  fake trace'],
         );
+      });
+      test('evaluates condition', () {
+        check(() => 1).isRejectedBy(
+          (it) => it.returnsNormally((it) => it.equals(2)),
+          actual: ['<1>'],
+          which: ['is not equal'],
+        );
+      });
+      test('returns valid subject', () {
+        check(() => 1).returnsNormally().equals(1);
       });
     });
   });